The logging time is now fixed

+++ code/system/application/knowledge/src/linguistics/v2vdenm/v2vdenm.cpp
@@ -276,11 +276,11 @@ void V2vDenm::tearDown()
 }
 
 
-uint32_t V2vDenm::GenerateGenerationTime() const
+uint64_t V2vDenm::GenerateGenerationTime() const
 {
   std::chrono::system_clock::time_point start2004TimePoint = 
       std::chrono::system_clock::from_time_t(m_timeType2004);
-  uint32_t millisecondsSince2004Epoch =
+  uint64_t millisecondsSince2004Epoch =
       std::chrono::system_clock::now().time_since_epoch() /
       std::chrono::milliseconds(1) 
       - start2004TimePoint.time_since_epoch() / std::chrono::milliseconds(1);
